extract vector_{encode, decode} functions to their own thing.
